Move language dropdown to project dashboard
