+++ package.json
@@ -42,6 +42,7 @@
     "mocha": "^3.2.0",
     "nib": "^1.1.2",
     "pui-react-buttons": "^8.1.0",
+    "pui-react-dropdowns": "^8.3.1",
     "pui-react-iconography": "^8.1.0",
     "pui-react-mixins": "^8.1.0",
     "pui-react-modals": "^8.1.0",
