--- src/styles/components/dropdown-menu.styl
@@ -1,7 +1,0 @@
-.dropdown .dropdown-menu ul
-  list-style none
-  margin 0
-  padding 0
-  max-height 10em
-  width 100%
-  overflow auto
